fix: add if condition on controlle file on delete function api

+++ backend/controllers/product.controller.js
@@ -54,11 +54,16 @@ export const updatedProducts = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
   //console.log('Deleting product with ID:', id); do test
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(400).json({ success: false, message: 'Invalid product ID' });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: 'Product deleted successfully' });
     } catch (error) {
         console.error('Error in Delete product:', error.message);
-        res.status(404).json({ success: false, message: 'Error 404: Product not found' });
+        res.status(500).json({ success: false, message: 'Error 500: Error server' });
     }
 }
